[level 2] Title: 점 찍기, Time: 0.20 ms, Memory: 79.3 MB -BaekjoonHub

+++ 프로그래머스/2/140107. 점 찍기/점 찍기.java
@@ -3,7 +3,7 @@ class Solution {
         long answer = 0;
         
         for(int x = 0; x <= d; x += k) {
-            long y = (long)Math.sqrt((long)Math.pow(d, 2) - (long)Math.pow(x, 2));
+            int y = (int)Math.sqrt((long)Math.pow(d, 2) - (long)Math.pow(x, 2));
             answer += (y / k) + 1;
         }
